Move cherry-pick files to correct path, update imports
